requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-secret-manager==2.16.3
google-auth==2.23.4
apscheduler==3.10.4
pydantic==2.4.2
python-multipart==0.0.6